refactor v2 operator metadata modules
